pyproject.toml
[project]
name = "signalavergecg"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biosppy>=2.2.4",
    "edflib-python>=1.0.8",
    "matplotlib>=3.10.8",
    "notebook>=7.5.4",
    "pandas>=3.0.1",
    "peakutils>=1.3.5",
    "scipy>=1.17.1",
]

[dependency-groups]
build = [
    "wheel>=0.46.3",
]
dev = [
    "flake8>=7.3.0",
    "poethepoet>=0.42.1",
    "pycodestyle>=2.14.0",
    "ruff>=0.15.4",
]
